# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateStudyRequest:
    properties:
      description:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - title
    type: object
  api.ErrorResponse:
    properties:
      code:
        type: string
      details:
        description: Only included in debug mode
        type: string
      message:
        type: string
    type: object
...
  api.StudyListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      studies:
        items:
          $ref: '#/definitions/domain.Study'
        type: array
      total:
        type: integer
    type: object
...
  api.UpdateStudyRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
  domain.Study:
    properties:
      case_count:
        description: Denormalized counters for efficient queries
        type: integer
      complete_raters:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/domain.StudyStatus'
      title:
        type: string
      total_responses:
        type: integer
      unique_raters:
        type: integer
      updated_at:
        type: string
    type: object
...
host: api.anklyze.es
info:
  contact:
    name: API Support
    url: https://github.com/jferrl/anklyze
  description: Ankle fracture classification API. Classifies fractures according to
    Danis-Weber, Lauge-Hansen, AO/OTA, and Bartonicek systems.
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/MIT
  title: Anklyze API
  version: "1.0"
paths:
  /api/admin/studies:
    get:
      parameters:
      - description: Filter by status (draft, active, closed)
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StudyListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List studies
      tags:
      - Admin Studies
    post:
      consumes:
      - application/json
      parameters:
      - description: Study details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateStudyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Study'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new study
      tags:
      - Admin Studies
  /api/admin/studies/{id}:
    delete:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete a study
      tags:
      - Admin Studies
    get:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StudyDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get study details
      tags:
      - Admin Studies
    put:
      consumes:
      - application/json
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateStudyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Study'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/activate:
    put:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Study'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Activate a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/cases:
    post:
      consumes:
      - application/json
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: Case details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AddCaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Case'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Add a case to a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/cases/{caseId}:
    delete:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: Case ID
        in: path
        name: caseId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Remove a case from a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/cases/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: Ordered list of case IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ReorderCasesRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Reorder cases in a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/close:
    put:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Study'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Close a study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/progress:
    get:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RaterProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get rater progress
      tags:
      - Admin Studies
  /api/admin/studies/{id}/raters:
    get:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StudyRatersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List study raters
      tags:
      - Admin Studies
    post:
      consumes:
      - application/json
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AddStudyRaterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.StudyRater'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Add rater to study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/raters/{userId}:
    delete:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Remove rater from study
      tags:
      - Admin Studies
  /api/admin/studies/{id}/reliability:
    get:
      parameters:
      - description: Study ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StudyReliabilityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get study reliability metrics
      tags:
      - Admin Studies
...
  /api/analytics/distribution/{system}:
    get:
      description: Returns detailed distribution for a specific classification system
      parameters:
      - description: Classification system (danis-weber, lauge-hansen, ao-ota)
        in: path
        name: system
        required: true
        type: string
      - default: 30 days ago
        description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - default: today
        description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Distribution data
          schema:
            $ref: '#/definitions/domain.ClassificationDistribution'
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get classification distribution
      tags:
      - Analytics
  /api/analytics/summary:
    get:
      description: Returns aggregated classification statistics for a time period
      parameters:
      - default: 30 days ago
        description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - default: today
        description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Analytics summary
          schema:
            $ref: '#/definitions/domain.AnalyticsSummary'
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get analytics summary
      tags:
      - Analytics
  /api/analytics/trends:
    get:
      description: Returns time-series classification data with configurable granularity
      parameters:
      - default: 30 days ago
        description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - default: today
        description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - default: day
        description: Time granularity (day, week, month)
        in: query
        name: granularity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trend data
          schema:
            $ref: '#/definitions/domain.TrendData'
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get classification trends
      tags:
      - Analytics
  /api/chat:
    post:
      consumes:
      - application/json
      description: Processes natural language fracture descriptions and returns classification
      parameters:
      - description: Chat message
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Chat response with classification
          schema:
            $ref: '#/definitions/service.ChatResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Chat service unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Chat-based fracture classification
      tags:
      - Chat
  /api/chat/session:
    post:
      description: Creates a new chat session for tracking conversations
      parameters:
      - default: en
        description: Language (en, es)
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session created
          schema:
            $ref: '#/definitions/api.CreateChatSessionResponse'
      summary: Create a new chat session
      tags:
      - Chat
  /api/chat/session/{id}/abandon:
    put:
      description: Marks a chat session as abandoned (user left before completion)
      parameters:
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Session abandoned
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark a chat session as abandoned
      tags:
      - Chat
  /api/chat/session/{id}/complete:
    put:
      description: Marks a chat session as successfully completed
      parameters:
      - description: Session ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Session completed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark a chat session as complete
      tags:
      - Chat
...
schemes:
- https
swagger: "2.0"